backend/requirements.txt
flask>=3.0
flask-cors>=4.0
PyJWT>=2.8
bcrypt>=4.1
python-dotenv>=1.0
